Move `Arch/i686/Memory.hpp` to `Arch/i686/Memory/`
